Move `db_models` to `src`
Incorrectly placed directory now moved under `src`
