+++ backend/tests/db_models/test_athlete.py
@@ -1,8 +1,8 @@
 from decimal import Decimal
 from unittest import TestCase
 
-from db_models.athlete import engine, Base as AthleteBase, Athlete
-from db_models.user import User, Base as UserBase
+from src.db_models.athlete import engine, Base as AthleteBase, Athlete
+from src.db_models.user import User, Base as UserBase
 from src import db
